move `getForeignCurrencyRates` to `google/sheets.js`
